create plot class and insert matplotlib code
