+++ src/core/sim.py
@@ -3,10 +3,9 @@ from core.config import Config
 from core.integrator.integrator import propagate_state
 from core.state import State, StateTime, ObservedState, PropagatedOutput
 from core.models.model_list import ModelContainer
-from utils.astropy_util import get_body_position
 from utils.log import log
-import matplotlib.pyplot as plt
-from utils.constants import BodyEnum, R_EARTH, R_MOON
+from utils.constants import R_EARTH
+from utils.matplotlib_util import Plot
 
 
 class CislunarSim:
@@ -23,20 +22,18 @@ class CislunarSim:
         self.should_run = True
         self.num_iters = 0
 
-        fig = plt.figure()
-        self.ax = fig.add_subplot(111, projection="3d")
-        self.ax.set_xlim(-10000000, 10000000)
-        self.ax.set_ylim(-10000000, 10000000)
-        self.ax.set_zlim(-10000000, 10000000)
-        self.ax.set_xlabel("x")
-        self.ax.set_ylabel("y")
-        self.ax.set_zlabel("z")
-        # self.ax = plt.axes(projection='3d')
+        # fig = plt.figure()
+        # self.ax = fig.add_subplot(111, projection="3d")
+        # self.ax.set_xlim(-10000000, 10000000)
+        # self.ax.set_ylim(-10000000, 10000000)
+        # self.ax.set_zlim(-10000000, 10000000)
+        # self.ax.set_xlabel("x")
+        # self.ax.set_ylabel("y")
+        # self.ax.set_zlabel("z")
+        # # self.ax = plt.axes(projection='3d')
         self.xlocs = np.array([])
         self.ylocs = np.array([])
         self.zlocs = np.array([])
-
-        self.ax.set
 
     def step(self) -> PropagatedOutput:
         """
@@ -97,28 +94,30 @@ class CislunarSim:
         if (state.x**2 + state.y**2 + state.z**2) ** 0.5 < R_EARTH:
             log.error("Stopping sim because craft is inside the Earth")
             log.debug(f"r={(state.x**2 + state.y**2 + state.z**2)**0.5} < {R_EARTH}")
-            self.plot_data()
+            data_plot = Plot(self.xlocs, self.ylocs, self.zlocs, self.state)
+            data_plot.plot_data()
+            # self.plot_data()
             return True
 
         return False
 
-    def plot_data(self) -> None:
-        """Procedure that plots a model of the earth, moon and the craft's trajectory in R3"""
-        # 3D scatter plot of craft's trajectory
-        self.ax.scatter3D(self.xlocs, self.ylocs, self.zlocs, cmap="Greens")
+    # def plot_data(self) -> None:
+    #     """Procedure that plots a model of the earth, moon and the craft's trajectory in R3"""
+    #     # 3D scatter plot of craft's trajectory
+    #     self.ax.scatter3D(self.xlocs, self.ylocs, self.zlocs, cmap="Greens")
 
-        # Calculation and plotting of earth's position
-        u, v = np.mgrid[0 : 2 * np.pi : 20j, 0 : np.pi : 10j]
-        earth_x = R_EARTH * np.cos(u) * np.sin(v)
-        earth_y = R_EARTH * np.sin(u) * np.sin(v)
-        earth_z = R_EARTH * np.cos(v)
-        self.ax.plot_surface(earth_x, earth_y, earth_z, color="g")
+    #     # Calculation and plotting of earth's position
+    #     u, v = np.mgrid[0 : 2 * np.pi : 20j, 0 : np.pi : 10j]
+    #     earth_x = R_EARTH * np.cos(u) * np.sin(v)
+    #     earth_y = R_EARTH * np.sin(u) * np.sin(v)
+    #     earth_z = R_EARTH * np.cos(v)
+    #     self.ax.plot_surface(earth_x, earth_y, earth_z, color="g")
 
-        # Calculation and plotting of moon's position
-        moon_cx, moon_cy, moon_cz = get_body_position(self.state.time, BodyEnum.Moon)
-        moon_x = moon_cx + R_MOON * np.cos(u) * np.sin(v)
-        moon_y = moon_cy + R_MOON * np.sin(u) * np.sin(v)
-        moon_z = moon_cz + R_MOON * np.cos(v)
-        self.ax.plot_surface(moon_x, moon_y, moon_z, color="g")
+    #     # Calculation and plotting of moon's position
+    #     moon_cx, moon_cy, moon_cz = get_body_position(self.state.time, BodyEnum.Moon)
+    #     moon_x = moon_cx + R_MOON * np.cos(u) * np.sin(v)
+    #     moon_y = moon_cy + R_MOON * np.sin(u) * np.sin(v)
+    #     moon_z = moon_cz + R_MOON * np.cos(v)
+    #     self.ax.plot_surface(moon_x, moon_y, moon_z, color="g")
 
-        plt.show()
+    #     plt.show()
